requirements.txt
alembic==1.4.2
altgraph==0.17
amqp==5.0.5
aniso8601==8.0.0
asgiref==3.2.10
beautifulsoup4==4.9.1
bidict==0.21.2
billiard==3.6.3.0
blinker==1.4
blueprint==3.4.2
bs4==0.0.1
cachelib==0.1.1
celery==5.0.5
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cycler==0.10.0
dnspython==1.16.0
dominate==2.5.2
eventlet==0.30.0
Faker==4.5.0
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.9
Flask-DebugToolbar==0.11.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
flask-redis==0.4.0
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-Session==0.3.2
Flask-SocketIO==5.0.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
future==0.18.2
game==0.0.5
greenlet==0.4.17
gunicorn==20.0.4
html5lib==1.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
kombu==5.0.2
lxml==4.6.2
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.19.5
pefile==2019.4.18
Pillow==8.1.0
prompt-toolkit==3.0.16
PyExecJS==1.5.1
PyJWT==1.7.1
PyMySQL==0.10.1
pyodbc==4.0.30
pyparsing==2.4.7
pypiwin32==223
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==4.0.0
python-socketio==5.0.2
pytz==2020.1
pywin32==300
pywin32-ctypes==0.2.0
redis==3.5.3
requests==2.24.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.19
sqlparse==0.3.1
text-unidecode==1.3
tornado==6.0.4
urllib3==1.25.10
vine==5.0.0
visitor==0.1.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
WTForms==2.3.3
wxPython==4.1.1
xlrd==2.0.1
